Split RPC into multiple files
