+++ crates/cheatnet/src/execution/mod.rs
@@ -1,0 +1,9 @@
+#[allow(clippy::module_name_repetitions)]
+pub mod cairo1_execution;
+pub mod contract_print;
+pub mod entry_point;
+pub mod events;
+#[allow(clippy::module_name_repetitions)]
+pub mod execution_info;
+pub mod hint_processor;
+pub mod syscalls;
