+++ crates/cheatnet/src/lib.rs
@@ -7,8 +7,8 @@ use state::{CheatcodeState, DictStateReader};
 
 pub mod cheatcodes;
 pub mod constants;
-pub mod contract_print;
 pub mod conversions;
+pub mod execution;
 pub mod panic_data;
 pub mod rpc;
 pub mod state;
